+++ src/web/webapi.go
@@ -14,7 +14,6 @@ import (
 	"encoding/gob"
 	"encoding/hex"
 	"encoding/json"
-	"flag"
 	"fmt"
 	"io"
 	"log"
@@ -156,14 +155,4 @@ func StartApiService() {
 	if err := http.ListenAndServe(":9000", nil); err != nil {
 		log.Fatal("ListenAndServe: ", err)
 	}
-}
-
-// Start Wasm
-func StartWebAseembly() {
-	listen := flag.String("listen", ":9000", "listen address")
-	dir := flag.String("dir", "./wasm", "directory to serve")
-	flag.Parse()
-
-	log.Printf("listening on %q...", *listen)
-	log.Fatal(http.ListenAndServe(*listen, http.FileServer(http.Dir(*dir))))
 }
